+++ app/Http/Controllers/API/UserController.php
@@ -17,6 +17,8 @@ class UserController extends BaseController
      */
     public function index(Request $request)
     {
+        if (!check_permission('READ', 'users'))
+            return $this->sendError('Forbidden', 'Page unaccessable', 403);
         $users = empty($request->all()) ? User::all() : User::where('type', (int)$request->query('type'))->get();
         return $this->sendResponse($users, 'Users retrieved successfully');
     }
@@ -29,6 +31,8 @@ class UserController extends BaseController
      */
     public function store(Request $request)
     {
+        if (!check_permission('CREATE', 'users'))
+            return $this->sendError('Forbidden', 'Page unaccessable', 403);
         $data = $request->all();
         $validator = Validator::make($data, [
             'name' => 'required',
@@ -52,6 +56,8 @@ class UserController extends BaseController
      */
     public function show(User $user)
     {
+        if (!check_permission('READ', 'users'))
+            return $this->sendError('Forbidden', 'Page unaccessable', 403);
         return $this->sendResponse(new UserResource($user), 'User retrieved successfully.');
     }
 
@@ -64,6 +70,8 @@ class UserController extends BaseController
      */
     public function update(Request $request, User $user)
     {
+        if (!check_permission('UPDATE', 'users'))
+            return $this->sendError('Forbidden', 'Page unaccessable', 403);
         $user->update($request->all());
         return $this->sendResponse(new UserResource($user), 'User updated successfully.');
     }
@@ -76,6 +84,8 @@ class UserController extends BaseController
      */
     public function destroy(User $user)
     {
+        if (!check_permission('DELETE', 'users'))
+            return $this->sendError('Forbidden', 'Page unaccessable', 403);
         $user->delete();
         return $this->sendResponse($user, 'User deleted successfully.');
     }
